:boom: 0.1.0
Incluído o sistema de notas e o sistema de segurança contra ataques CSRF.

+++ dbconnect/delete_author.php
@@ -1,31 +1,38 @@
 <?php
 require "../config.php";
 
-$id = $_POST['id'];
+$data = formTreatment($_POST);
+$id = $data['id_author'];
 
-if (!empty($id)) {
-    $request_item_to_delete = $conn->prepare("SELECT * FROM author WHERE id_author = ?");
-    $request_item_to_delete->bind_param('i', $id);
-    $request_item_to_delete->execute();
-    $result = $request_item_to_delete->get_result();
-    $request_item_to_delete->close();
-    $author = $result->fetch_assoc();
-    if($author) {
-        $query = $conn->prepare("DELETE FROM author WHERE id_author = ?");
-        $query->bind_param('i', $id);
+if ($data['csrf_token'] == $_SESSION['csrf_token']) {
+    if (!empty($id)) {
+        $request_item_to_delete = $conn->prepare("SELECT * FROM author WHERE id_author = ?");
+        $request_item_to_delete->bind_param('i', $id);
+        $request_item_to_delete->execute();
+        $result = $request_item_to_delete->get_result();
+        $request_item_to_delete->close();
+        $author = $result->fetch_assoc();
+        if($author) {
+            $query = $conn->prepare("DELETE FROM author WHERE id_author = ?");
+            $query->bind_param('i', $id);
 
-        if ($query->execute() === TRUE) {
-            $message = $author['name']." excluído(a) com sucesso.";
-            header("Location: /PIE3/pages/author/authors.php?success=".urlencode($message));
-            exit();
+            if ($query->execute() === TRUE) {
+                $message = $author['name']." excluído(a) com sucesso.";
+                header("Location: /PIE3/pages/author/authors.php?success=".urlencode($message));
+                exit();
+            } else {
+                $message = "O seguinte erro ocorreu ao tentar excluir o autor ".$author['name'].":<br>".$query->error;
+                header("Location: /PIE3/pages/author/authors.php?error=".urlencode($message));
+                exit();
+            }
         } else {
-            $message = "O seguinte erro ocorreu ao tentar excluir o autor ".$author['name'].":<br>".$query->error;
-            header("Location: /PIE3/pages/author/authors.php?error=".urlencode($message));
+            header("Location: /PIE3/pages/author/authors.php?error=Registro não encontrado para a chave enviada.");
             exit();
         }
-    } else {
-        header("Location: /PIE3/pages/author/authors.php?error=Registro não encontrado para a chave enviada.");
-        exit();
+        
     }
-    
+} else {
+    $message = "Token de segurança inválido.";
+    header("Location: /PIE3/pages/publisher/publishers.php?error=".urlencode($message));
+    exit();
 }
